Fix for dijit/form/Select and dijit/form/FilteringSelect
When using those widget, it's better to use the item field instead of value
to fit the angularjs way of doing things.

Like that, the select returns the full object in the model and not the just the id.

Now when the model changes, the dojo widget is also updated automatically.